requirements.txt
matplotlib==1.5.1
nltk==3.1
numpy==1.10.4
PyYAML==3.11
scipy==0.17.0
urwid==1.3.1
